Add /langs endpoint
I thought about adding the langs to the return value of the the card
search endpoint, but it would complicate our model for the `Cards`
because the list of all langs doesn't come for free in the JSON dumps
from scryfall which would lead to a more complicated import process.

It also feels kind of weird because a "Card" doesn't have multiple
languages. It doesn't make a lot of sense for a card that includes a
link to an image with Korean text to have a list of languages to me.
Perhaps there's a solution where we have multiple types where one is a
specific instance of a card and the other is a class of cards.

There's a chance I'll want to add this data later, but for now an extra
call to get the langs when needed is fine.